requirements.txt
asgiref==3.2.7
autopep8==1.5
bson==0.5.8
coverage==5.0.4
dataclasses==0.6
Django==2.2.12
django-appconf==1.0.4
djangorestframework==3.11.0
djongo==1.3.1
gunicorn==20.0.4
pycodestyle==2.5.0
pymongo==3.10.1
pysqlite3==0.4.2
python-dateutil==2.8.1
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
selenium==3.141.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.25.8
